+++ client/src/App.js
@@ -14,7 +14,7 @@ import Cart from "./components/cartComponent/cart";
 import { AuthProvider } from "./components/context/Auth-context.js";
 // import { ProtectedRoutes } from "./components/context/ProtectedRoutes";
 import Profile from "./components/pages/profile/profile";
-import EditCourse from "./components/editcourse/EditCourse2";
+import EditCourse from "./components/editcourse/EditCourse";
 import Create from "./components/create/Create";
 import ShowMoreCourses from "./components/pages/profile/ShowMoreCourses";
 
@@ -27,26 +27,31 @@ import "react-toastify/dist/ReactToastify.css";
 function App() {
   return (
     <AuthProvider>
-    <PayPalScriptProvider options={{ "client-id" : "ARjP3IEG_UlE_vM5S-HxokNsAdUIkcck3Jr_O3cS3WLaL0sD85BCGoe5-FG2qbOj9DBy0R4RdTdD-BlD"}}>
-    <ToastContainer/>
-    <ChakraProvider>
-      <Route exact path="/" component={Home} />
-      <Route exact path="/blog" component={Blog} />
-      <Route path="/cart" component={Cart} />
-      <Route exact path="/contact" component={Contact} />
-      <Route exact path="/signup" component={Signup} />
-      <Route exact path="/login" component={Login} />
-      <Route exact path="/course" component={Course} />
-      <Route exact path="/about" component={About} />
-      <Route exact path="/detail/:id" component={Detail} />
-      <Route exact path="/detailVideo/:courseId/:id" component={Videos} />
-      <Route exact path="/profile" component={Profile} />
-      <Route exact path="/editcourse/:courseId" component={EditCourse}/>
-      
-      <Route exact path="/formCourse" component={Create} />
-      <Route exact path="/cursosCreados" component={ShowMoreCourses} />
-    </ChakraProvider>
-    </PayPalScriptProvider>
+      <PayPalScriptProvider
+        options={{
+          "client-id":
+            "ARjP3IEG_UlE_vM5S-HxokNsAdUIkcck3Jr_O3cS3WLaL0sD85BCGoe5-FG2qbOj9DBy0R4RdTdD-BlD",
+        }}
+      >
+        <ToastContainer />
+        <ChakraProvider>
+          <Route exact path="/" component={Home} />
+          <Route exact path="/blog" component={Blog} />
+          <Route path="/cart" component={Cart} />
+          <Route exact path="/contact" component={Contact} />
+          <Route exact path="/signup" component={Signup} />
+          <Route exact path="/login" component={Login} />
+          <Route exact path="/course" component={Course} />
+          <Route exact path="/about" component={About} />
+          <Route exact path="/detail/:id" component={Detail} />
+          <Route exact path="/detailVideo/:courseId/:id" component={Videos} />
+          <Route exact path="/profile" component={Profile} />
+          <Route exact path="/editcourse/:courseId" component={EditCourse} />
+
+          <Route exact path="/formCourse" component={Create} />
+          <Route exact path="/cursosCreados" component={ShowMoreCourses} />
+        </ChakraProvider>
+      </PayPalScriptProvider>
     </AuthProvider>
   );
 }
