+++ client/src/components/create/Create.jsx
@@ -5,32 +5,12 @@ import { getCategories } from "../../redux/actions/index"
 import Navbar from "../navbar/Navbar.jsx"
 import Footer2 from "../footer/Footer2.jsx"
 import {
-  FormControl,
-  FormLabel,
-  Input,
-  Box,
-  Button,
-  Text,
-  Alert,
-  AlertIcon,
-  HStack,
-  Tag,
-  TagLabel,
-  Tooltip,
-  Select,
+  FormControl, FormLabel, Input, Box, Button, Text, Alert, AlertIcon, HStack, Tag, TagLabel, Tooltip, Select,
 } from '@chakra-ui/react'
 import { ArrowBackIcon } from '@chakra-ui/icons'
 import axios from "axios"
 import { validate, validateVideo } from "./validate";
 
-//FALTAN VALIDACIONES EN EL FORMULARIO DEL VIDEO
-//FALTAN VALIDACIONES EN EL FORMULARIO DEL VIDEO
-//FALTAN VALIDACIONES EN EL FORMULARIO DEL VIDEO
-//FALTAN VALIDACIONES EN EL FORMULARIO DEL VIDEO
-//FALTAN VALIDACIONES EN EL FORMULARIO DEL VIDEO
-//FALTAN VALIDACIONES EN EL FORMULARIO DEL VIDEO
-//FALTAN VALIDACIONES EN EL FORMULARIO DEL VIDEO
-//FALTAN VALIDACIONES EN EL FORMULARIO DEL VIDEO
 const Create = () => {
   const dispatch = useDispatch();
   const history = useHistory()
@@ -48,16 +28,7 @@ const Create = () => {
     price: null,
   })
 
-
-  const [inputVideo, setInputVideo] = useState({
-    nameVideo: '',
-    urlVideo: "",
-    description: "",
-  })
-
   const [errors, setErrors] = useState({})
-
-  const [errorsVideo, setErrorsVideo] = useState({})
 
   const handelChange = (e) => {
     if (e.target.name === "Thumbnail") {
@@ -77,13 +48,6 @@ const Create = () => {
         [e.target.name]: e.target.value
       })
     }
-  }
-
-  const handelChangeVideo = (e) => {
-    setInputVideo({
-      ...inputVideo,
-      [e.target.name]: e.target.value
-    })
   }
 
   const handelSubmit = async () => {
@@ -107,19 +71,6 @@ const Create = () => {
     history.push(`/editcourse/${response.data.course.id}`)
   }
 
-  // const handelSubmitVideo = () => {
-  //   setInput({
-  //     ...input,
-  //     video: [...input.video, inputVideo]
-  //   })
-  //   setInputVideo({
-  //     nameVideo: '',
-  //     urlVideo: "",
-  //     description: "",
-  //   })
-  //   document.getElementById("formVideo").reset()
-  // }
-
   const handelDelete = (e) => {
     const name = e.target.name
     const index = input.category.findIndex(e => e === name)
@@ -131,22 +82,11 @@ const Create = () => {
     })
   }
 
-  // const handelDeleteVideo = (name) => {
-  //   const index = input.video.findIndex(e => e.name === name)
-  //   let videos = input.video
-  //   videos.splice(index, 1)
-  //   setInput({
-  //     ...input,
-  //     video: videos
-  //   })
-  // }
-
   useEffect(() => {
     if (!categories.length)
       dispatch(getCategories())
     setErrors(validate(input))
-    setErrorsVideo(validateVideo(inputVideo, input))
-  }, [dispatch, input, inputVideo])
+  }, [dispatch, input])
 
 
 
@@ -208,56 +148,13 @@ const Create = () => {
                 <AlertIcon />
                 {errors.price}
               </Alert>}
-              {/* <Box border="1px" borderRadius="20" p="10" my="10">
-                <FormControl id="formVideo">
-                  <FormLabel display="flex" justifyContent="center">Video:</FormLabel>
-                  <FormLabel>Name:</FormLabel>
-                  <Input name="nameVideo" onChange={handelChangeVideo} />
-                  {errorsVideo.nameVideo && <Text fontSize='10px' color='red'>
-                    {errorsVideo.nameVideo}
-                  </Text>}
-                  <FormLabel>Description:</FormLabel>
-                  <Input name="description" onChange={handelChangeVideo} />
-                  {errorsVideo.description && <Text fontSize='10px' color='red'>
-                    {errorsVideo.description}
-                  </Text>}
-                  <FormLabel>urlVideo:</FormLabel>
-                  <Input name="urlVideo" onChange={handelChangeVideo} />
-                  {errorsVideo.urlVideo && <Text fontSize='10px' color='red'>
-                    {errorsVideo.urlVideo}
-                  </Text>}
-                  <Box display="flex" flexDirection="row-reverse">
-                    <Button mt={4} colorScheme='teal' onClick={handelSubmitVideo} disabled={Object.keys(errorsVideo).length ? true : false}>
-                      Add Video
-                    </Button>
-                  </Box>
-                </FormControl>
-
-              </Box> */}
-              <Text textAlign="center">Todos estos datos se podran modificar luego de crear el curso.</Text>
+              <Text textAlign="center">All these data can be modified after creating the course.</Text>
               <Button mt={4} colorScheme='teal' disabled={Object.keys(errors).length ? true : false} onClick={handelSubmit}>
-                Submit
+                Create
               </Button>
             </FormControl>
           </Box>
         </Box >
-        {/* <Box w='65%' my="5" display='flex' flexDirection='column' alignItems='center' justifyContent='center' h='-webkit-fit-content' >
-          <Text fontSize='3xl'>List Videos</Text>
-          <Box border='1px' display='grid' gridTemplateColumns='repeat(2, 1fr)' gridAutoRows='200px' justifyContent='center' w='95%' borderRadius='25' py='5' gap='10'>
-
-            {
-              input.video?.map((e, i) => (
-                <Box w='90%' h='200' bg='gray.100' key={i} borderRadius='20' display='flex' flexDirection='column' alignItems='center' justifyContent='center' ml='5' textAlign='center' whiteSpace='initial' overflow='hidden'>
-                  <h1>{e.name}</h1>
-                  <h4>{e.description}</h4>
-                  <h6>{e.urlVideo}</h6>
-                  <Button colorScheme='red' m='2' onClick={() => handelDeleteVideo(e.name)}> X </Button>
-                </Box>
-              ))
-            }
-
-          </Box>
-        </Box> */}
       </Box >
       <Footer2 />
     </Box>
