+++ client/src/components/create/validate.js
@@ -1,14 +1,16 @@
 export function validate(input) {
   const errors = {};
-  if (input.name.length < 3)
+  if (input.name?.length <= 3)
     errors.name = "The name must contain more than 3 letters";
-  if (input.description.length < 3)
+  if (input.description?.length <= 3)
     errors.description = "The description must contain more than 3 letters";
-  if (input.category.length < 1)
+  if (input.category?.length < 1)
     errors.category = "It is obligatory to add at least one category";
-  if (input.price === null || input.price === "")
+  if (input.categories?.length < 1)
+    errors.category = "It is obligatory to add at least one category";
+  if (input?.price === null || input?.price === "")
     errors.price = "You must enter a price";
-  else if (input.price < 1) errors.price = "It cannot be a negative price";
+  else if (input?.price < 1) errors.price = "It cannot be a negative price";
   return errors;
 }
 
@@ -24,5 +26,6 @@ export function validateVideo(inputVideo, inputCourse) {
     errors.urlVideo = "The url must have .com";
   if (inputVideo.description?.length <= 3)
     errors.description = "A description for the video is required";
+
   return errors;
 }
