+++ client/src/components/home/Home.jsx
@@ -1,7 +1,6 @@
 import React, { useState } from "react";
 import Navbar from '../navbar/Navbar';
 import style from "./Home.module.css"
-// import SearchBar from "../searchbar/SearchBar.jsx";
 import CourseCard from '../card/CourseCard.jsx';
 import { Heading, Box, Text, Image } from '@chakra-ui/react'
 import mainpicture from "../../image/maintextimage.png"
@@ -18,130 +17,115 @@ import CreateReviewPage from "../createReviewPage/CreateReviewPage";
 
 export default function Home() {
 
-    // carrousel: 
-    let courses = useSelector(state => state.courses);
-    courses = courses?.sort((a, b) => b.rating - a.rating)?.slice(0, 15)
-    const [coursesPerPage] = useState(3);
-    const [currentPage, setCurrentPage] = useState(1);
-    const last = currentPage * coursesPerPage;
-    const first = last - coursesPerPage;
-    const currentCourses = courses?.slice(first, last);
-    const numOfPages = courses.length / coursesPerPage;
+  let courses = useSelector(state => state.courses);
+  courses = courses?.sort((a, b) => b.rating - a.rating)?.slice(0, 15)
 
-    const handleNext = (e) => {
-        e.preventDefault();
-        currentPage < numOfPages ? setCurrentPage(currentPage + 1) : setCurrentPage(1);
-    }
+  const [coursesPerPage] = useState(3);
+  const [currentPage, setCurrentPage] = useState(1);
+  const last = currentPage * coursesPerPage;
+  const first = last - coursesPerPage;
+  const currentCourses = courses?.slice(first, last);
+  const numOfPages = courses.length / coursesPerPage;
 
-    const handlePrevious = (e) => {
-        e.preventDefault();
-        currentPage > 1 ? setCurrentPage(currentPage - 1) : setCurrentPage(1);
-    }
+  const handleNext = (e) => {
+    e.preventDefault();
+    currentPage < numOfPages ? setCurrentPage(currentPage + 1) : setCurrentPage(1);
+  }
 
-    // cards:
-    const dispatch = useDispatch();
-    useEffect(() => {
-        // if (!courses.length) {
-        dispatch(getAllCourses())
-        // }
-    }, [dispatch]);
+  const handlePrevious = (e) => {
+    e.preventDefault();
+    currentPage > 1 ? setCurrentPage(currentPage - 1) : setCurrentPage(1);
+  }
 
+  const dispatch = useDispatch();
+  useEffect(() => {
+    dispatch(getAllCourses())
+  }, [dispatch]);
 
+  return (
 
-    // const ListHeader = ({children}) => {
-    //     return (
-    //         <Text fontWeight={'500'} fontSize={'lg'} mb={2}>
-    //             {children}
-    //         </Text>
-    //     );
-    // };
-
-    return (
+    <div>
+      <div className={style.bg}>
+        <Navbar>
+        </Navbar>
+      </div>
+      <div className={style.maintext}>
+        <Box maxW='32rem'>
+          <Text fontSize='xl'>
+            START TO SUCCESS
+          </Text>
+          <Heading mb={4}>Acces To 5000+ Courses from
+            300 Instructor & Institutions</Heading>
+          <Text fontSize='xl'> Varius version have envolved over the years, sometimes by accident,
+          </Text>
+        </Box>
 
         <div>
-            <div className={style.bg}>
-                <Navbar>
-                </Navbar>
-            </div>
-            <div className={style.maintext}>
-                <Box maxW='32rem'>
-                    <Text fontSize='xl'>
-                        START TO SUCCESS
-                    </Text>
-                    <Heading mb={4}>Acces To 5000+ Courses from
-                        300 Instructor & Institutions</Heading>
-                    <Text fontSize='xl'> Varius version have envolved over the years, sometimes by accident,
-                    </Text>
-                    {/* <SearchBar>
-                    </SearchBar> */}
-                </Box>
-
-                <div>
-                    <Box boxSize='m'>
-                        <Image src={mainpicture} alt='image' />
-                    </Box>
-
-                </div>
-            </div>
-
-            <br>
-            </br>
-
-            <div title="carousel">
-                {/* <Carousel>
-
-                </Carousel> */}
-                <div className={style.coursecont}>
-                    <Icon as={ArrowLeftIcon} onClick={(e) => handlePrevious(e)}
-                        className={style.icon} />
-                    {
-                        currentCourses.map((course) => {
-                            return (
-                                <CourseCard
-                                    key={course.id}
-                                    name={course.name}
-                                    teacherId={course.teacherId}
-                                    teacherName={course.teacherName}
-                                    id={course.id}
-                                    // description={course.description}
-                                    price={course.price}
-                                    rating={course.rating}
-                                    categories={course.categories}
-                                    image={course.image}
-                                />)
-                        })
-                    }
-                    <Icon as={ArrowRightIcon} onClick={(e) => handleNext(e)}
-                        className={style.icon} />
-                </div>
-            </div>
-            <br />
-
-            <div>
-                <Categories />
-
-            </div>
-
-            <div>
-                <BasicStatistics>
-                </BasicStatistics>
-            </div>
-
-            <div>
-                <Hola>
-                </Hola>
-            </div>
-            
-            <div>
-                <CreateReviewPage/>
-            </div>
-
-            <div>
-                <Footer2>
-                </Footer2>
-            </div>
+          <Box boxSize='m'>
+            <Image src={mainpicture} alt='image' />
+          </Box>
 
         </div>
+      </div>
 
-    );
+      <br>
+      </br>
+
+      <div title="carousel">
+        <div className={style.coursecont}>
+          <Icon as={ArrowLeftIcon} onClick={(e) => handlePrevious(e)}
+            className={style.icon} />
+          {
+            currentCourses.map((course) => {
+              if (course.archieved === false && course.status === "APPROVED") {
+                return (
+                  <CourseCard
+                    key={course.id}
+                    name={course.name}
+                    teacherId={course.teacherId}
+                    teacherName={course.teacherName}
+                    id={course.id}
+                    // description={course.description}
+                    price={course.price}
+                    rating={course.rating}
+                    categories={course.categories}
+                    image={course.image}
+                    videos={course.videos}
+                    archieved={course.archieved}
+                    status={course.status}
+                  />)
+              }
+            })
+          }
+          <Icon as={ArrowRightIcon} onClick={(e) => handleNext(e)}
+            className={style.icon} />
+        </div>
+      </div>
+      <br />
+
+      <div>
+        <Categories />
+
+      </div>
+
+      <div>
+        <BasicStatistics />
+      </div>
+
+      <div>
+        <Hola />
+      </div>
+
+      <div>
+        <CreateReviewPage />
+      </div>
+
+      <div>
+        <Footer2>
+        </Footer2>
+      </div>
+
+    </div>
+
+  );
 }
