+++ client/src/components/pages/profile/profile.jsx
@@ -51,7 +51,7 @@ const Profile = () => {
             <GridItem gap='20' colSpan={2} display="flex" flexDirection="row" alignItems="center" justifyContent='center' mt="5" mb='5'>
               {
                 coursesCreateUser.length ? coursesCreateUser?.map(e => (
-                  <CourseCard key={e.id} id={e.id} teacherName={e.teacherName} name={e.name} description={e.description} rating={e.rating} price={e.price} image={e.image} categories={e.categories} />
+                  <CourseCard key={e.id} id={e.id} teacherName={e.teacherName} name={e.name} description={e.description} rating={e.rating} price={e.price} image={e.image} categories={e.categories} archieved={e.archieved} status={e.status} />
                 ))
                   : undefined
               }
